+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [3.1.10] - 2022-01-10
 ### Fixed
 - Issues with Group Polls [#1157](https://github.com/rokwire/illinois-app/issues/1157).
+### Changed
+- Cleanup and fixes for Auth2 login and refresh token [#1168](https://github.com/rokwire/illinois-app/issues/1168).
 
 ## [3.1.9] - 2022-01-07
 ### Fixed